requirements.txt
absl-py==0.9.0
tensorflow==1.14.0
beautifulsoup4==4.8.2